+++ CHANGELOG.md
@@ -16,6 +16,7 @@ through rest.
 * Added `User#flag`.
 * Added `GuildFeature.WelcomeScreenEnabled`.
 * Added `MessageModifyBuilder#allowedMentions`.
+* `Embeds` can now copy their contents over to builders by using `Embed#apply(EmbedBuilder)`.
 
 ## Fixes